+++ packages/components-vue/lib/components.ts
@@ -22,15 +22,25 @@ export const IfxAccordionItem = /*@__PURE__*/ defineContainer<JSX.IfxAccordionIt
 
 export const IfxAlert = /*@__PURE__*/ defineContainer<JSX.IfxAlert>('ifx-alert', undefined, [
   'color',
-  'icon'
+  'icon',
+  'ifxClose'
 ]);
 
 
 export const IfxBadge = /*@__PURE__*/ defineContainer<JSX.IfxBadge>('ifx-badge', undefined);
 
 
-export const IfxBreadcrumb = /*@__PURE__*/ defineContainer<JSX.IfxBreadcrumb>('ifx-breadcrumb', undefined, [
-  'breadcrumbs'
+export const IfxBreadcrumb = /*@__PURE__*/ defineContainer<JSX.IfxBreadcrumb>('ifx-breadcrumb', undefined);
+
+
+export const IfxBreadcrumbItem = /*@__PURE__*/ defineContainer<JSX.IfxBreadcrumbItem>('ifx-breadcrumb-item', undefined);
+
+
+export const IfxBreadcrumbItemLabel = /*@__PURE__*/ defineContainer<JSX.IfxBreadcrumbItemLabel>('ifx-breadcrumb-item-label', undefined, [
+  'icon',
+  'url',
+  'target',
+  'breadcrumbMenuIconWrapper'
 ]);
 
 
@@ -104,6 +114,8 @@ export const IfxDropdownItem = /*@__PURE__*/ defineContainer<JSX.IfxDropdownItem
   'icon',
   'checkable',
   'value',
+  'url',
+  'target',
   'itemValues'
 ]);