+++ packages/components/CHANGELOG.md
@@ -1,3 +1,34 @@
+# v18.2.3 (Fri Jun 16 2023)
+
+#### 🐛 Bug Fix
+
+- 159 restructuring breadcrumb [#172](https://github.com/Infineon/infineon-design-system-stencil/pull/172) ([@tishoyanchev](https://github.com/tishoyanchev))
+
+#### Authors: 1
+
+- Tisho Yanchev ([@tishoyanchev](https://github.com/tishoyanchev))
+
+---
+
+# v18.2.2 (Wed Jun 14 2023)
+
+#### 🐛 Bug Fix
+
+- Alert close event [#194](https://github.com/Infineon/infineon-design-system-stencil/pull/194) ([@tishoyanchev](https://github.com/tishoyanchev))
+
+#### 🔩 Dependency Updates
+
+- Bump vite from 4.3.8 to 4.3.9 in /examples/stencil-components/vanilla-cdn [#169](https://github.com/Infineon/infineon-design-system-stencil/pull/169) ([@dependabot[bot]](https://github.com/dependabot[bot]))
+- Bump vite from 4.3.8 to 4.3.9 in /examples/wrapper-components/react-typescript [#168](https://github.com/Infineon/infineon-design-system-stencil/pull/168) ([@dependabot[bot]](https://github.com/dependabot[bot]))
+- Bump vite from 4.3.8 to 4.3.9 in /examples/wrapper-components/vue-javascript [#167](https://github.com/Infineon/infineon-design-system-stencil/pull/167) ([@dependabot[bot]](https://github.com/dependabot[bot]))
+
+#### Authors: 2
+
+- [@dependabot[bot]](https://github.com/dependabot[bot])
+- Tisho Yanchev ([@tishoyanchev](https://github.com/tishoyanchev))
+
+---
+
 # v18.2.1 (Tue Jun 13 2023)
 
 #### 🐛 Bug Fix